# api_docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KrishiSetu AI Backend",
    "description": "Smart India Hackathon 2026 (SIH 26033) - Feature Phone First Farmer Marketplace",
    "version": "1.0.0"
  },
  "paths": {
    "/voice/incoming": {
      "post": {
        "tags": [
          "Voice Hotline"
        ],
        "summary": "Incoming Call",
        "description": "Simulates incoming phone call from farmer on 1800-260-3300 hotline.",
        "operationId": "incoming_call_voice_incoming_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceIncomingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceStepResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/voice/step": {
      "post": {
        "tags": [
          "Voice Hotline"
        ],
        "summary": "Voice Step",
        "description": "Processes each conversational step in the voice hotline.",
        "operationId": "voice_step_voice_step_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceStepRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceStepResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/voice/transcribe": {
      "post": {
        "tags": [
          "Voice Hotline"
        ],
        "summary": "Transcribe Voice",
        "description": "Speech-to-Text simulation (Whisper / AI4Bharat adapter).",
        "operationId": "transcribe_voice_voice_transcribe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceTranscribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceTranscribeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VoiceIncomingRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "title": "Phone",
            "default": "+919812345001"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "hi"
          }
        },
        "type": "object",
        "title": "VoiceIncomingRequest"
      },
      "VoiceStepRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "title": "Phone",
            "default": "+919812345001"
          },
          "step": {
            "type": "integer",
            "title": "Step",
            "default": 0
          },
          "user_speech": {
            "type": "string",
            "title": "User Speech",
            "default": ""
          },
          "session_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Data"
          }
        },
        "type": "object",
        "title": "VoiceStepRequest"
      },
      "VoiceStepResponse": {
        "properties": {
          "step": {
            "type": "integer",
            "title": "Step"
          },
          "ai_speech": {
            "type": "string",
            "title": "Ai Speech"
          },
          "audio_cue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Cue"
          },
          "is_final": {
            "type": "boolean",
            "title": "Is Final",
            "default": false
          },
          "extracted_entities": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted Entities"
          },
          "created_listing_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created Listing Id"
          },
          "session_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Session Data"
          }
        },
        "type": "object",
        "required": [
          "step",
          "ai_speech",
          "session_data"
        ],
        "title": "VoiceStepResponse"
      },
      "VoiceTranscribeRequest": {
        "properties": {
          "mock_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mock Text"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "hi"
          }
        },
        "type": "object",
        "title": "VoiceTranscribeRequest"
      },
      "VoiceTranscribeResponse": {
        "properties": {
          "transcript": {
            "type": "string",
            "title": "Transcript"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "default": 0.98
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "hi"
          },
          "detected_entities": {
            "additionalProperties": true,
            "type": "object",
            "title": "Detected Entities"
          }
        },
        "type": "object",
        "required": [
          "transcript",
          "detected_entities"
        ],
        "title": "VoiceTranscribeResponse"
      }
    }
  }
}